dice are now always positive
Using negative numbers for black's dice only led to subtle bugs and a
lot of abs().

+++ src/gibbon-jelly-fish-writer.c
@@ -353,7 +353,8 @@ static gchar *
 gibbon_jelly_fish_writer_double (const GibbonJellyFishWriter *self,
                                   const GibbonPosition *position)
 {
-        return g_strdup_printf ("  Doubles => %u", position->cube << 1);
+        return g_strdup_printf ("  Doubles => %llu",
+                                (unsigned long long) position->cube << 1);
 }
 
 static gchar *
